requirements.txt
absl_py==1.4.0
apache_beam==2.47.0
contextlib2==21.6.0
gin==0.1.006
ipython==8.9.0
joblib==1.2.0
lvis==0.5.3
lxml==4.9.2
matplotlib==3.6.3
mock==5.0.2
nets==0.0.3.1
nltk==3.8.1
numpy==1.20.3
opencv_python==4.7.0.68
orbit==0.2
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
portpicker==1.5.2
protobuf==3.20.3
pycocotools==2.0.6
PyYAML==6.0
PyYAML==6.0
sacrebleu==2.3.1
scikit_learn==1.2.1
scipy==1.10.1
sentencepiece==0.1.99
seqeval==1.2.2
setuptools==58.1.0
six==1.15.0
streamlit==1.22.0
tensorflow==2.10.0
tensorflow_addons==0.20.0
tensorflow_datasets==4.9.2
tensorflow_hub==0.13.0
tensorflow_model_optimization==0.7.4
tensorflow_text==2.12.1
tf_slim==1.1.0
